controller/backend/requirements.txt
# VISHWAAS Master - Backend Dependencies
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0

# Database migrations
alembic>=1.13.0

# Validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client for agent calls
httpx>=0.27.0

# CORS and utilities
python-multipart>=0.0.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate limiting
slowapi>=0.1.9

# Structured JSON logging
python-json-logger>=2.0.7

# Prometheus metrics
prometheus-client>=0.20.0
